No idea how these didn't get renamed in one of the patches...
