+++ crypto/CMakeLists.txt
@@ -604,7 +604,7 @@ if(BUILD_TESTING)
     fipsmodule/rand/fork_detect_test.cc
     fipsmodule/service_indicator/service_indicator_test.cc
     fipsmodule/sha/sha_test.cc
-    fipsmodule/cpucap/cpu-arm-linux_test.cc
+    fipsmodule/cpucap/cpu_arm_linux_test.cc
     hkdf/hkdf_test.cc
     hpke/hpke_test.cc
     hmac_extra/hmac_test.cc
