+++ crypto/fipsmodule/cpucap/cpu_arm_linux_test.cc
@@ -12,7 +12,7 @@
  * OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF OR IN
  * CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE. */
 
-#include "cpu-arm-linux.h"
+#include "cpu_arm_linux.h"
 
 #include <string.h>
